Created book display function, modal dialog
Book display function loops through the myLibrary
array and creates a div dom element with the text
content being the bookinfo of the last book object,
also added a remove button to the dom element. Currently
the bookDisplay function is inside the addBookToLibrary
function, I may change that later.
Created a modal dialog with text inputs and a checkbox
that calls the addBookToLibrary function with those values.